chore(engine-v2): Split benchmarks & add new ones for the preparation phase
